pyproject.toml
[tool.poetry]
name = "hierarhical_data"
version = "0.1.0"
description = ""
authors = ["Nikal Morgan <nikalmorgan11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^2.2.9"

[tool.poetry.dev-dependencies]
django-mptt = "^0.11.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
